Finalise Interactivity feature, correct linter errors
