+++ src/index.html
@@ -15,7 +15,7 @@
     </h1>
   
     <div class="input-container">
-      <input class="addtaks" type="text" id="new" placeholder="Add to your list...">
+      <input class="addtasks" type="text" id="new" placeholder="Add to your list...">
       <span class="btn fa fa-plus" id="add-btn"></span>
     </div>
