+++ src/index.js
@@ -1,86 +1,14 @@
 import './style.css';
-import myDateTime from './date-time.js';
+import myDateTime from './modules/date-time.js';
+import {
+  finalTodo, storeItem, addItem, editItem, removeItem, findIndex, clearTasks,
+} from './modules/addEditRemove.js';
+import { checkedBox, notChecked } from './modules/taskCompleted.js';
 
 // Declare initial variables
 const list = document.querySelector('#list');
 const addBtn = document.querySelector('#add-btn');
 const clear = document.querySelector('#clear');
-
-// Variable to get Tasks from local storage
-let todo = JSON.parse(localStorage.getItem('items')) || [];
-
-// Function to save to local storage
-const storeItem = () => {
-  localStorage.setItem('items', JSON.stringify(todo));
-};
-
-// Add new tasks
-const addItem = (desc) => {
-  const item = {
-    desc,
-    completed: false,
-    index: todo.length + 1,
-  };
-  todo.push(item);
-  storeItem();
-};
-
-// Edit existing tasks
-const editItem = (index, desc) => {
-  todo[index].desc = desc;
-  storeItem();
-};
-
-// Remove tasks
-const removeItem = (index) => {
-  todo.splice(index, 1);
-  for (let i = index; i < todo.length; i + 1) {
-    todo[i].index = i + 1;
-  }
-  storeItem();
-};
-
-// Find task indices
-const findIndex = (e) => {
-  const items = document.querySelectorAll('.item');
-  let index = 0;
-
-  for (let i = 0; i < items.length; i += 1) {
-    if (e.target.textContent === todo[i].desc) {
-      index = i;
-    } if (e.target.nextSibling.textContent === todo[i].desc) {
-      index = i;
-    } else if (e.target.previousSibling.textContent === todo[i].desc) {
-      index = i;
-    }
-  }
-
-  return index;
-};
-
-// Function to clear marked completed tasks
-function clearTasks() {
-  const unchecked = todo.filter((item) => item.completed === false);
-  unchecked.forEach((item, index) => {
-    item.index = index;
-  });
-  todo = unchecked;
-  storeItem();
-}
-
-const finalTodo = todo;
-
-// Set checkbox to true and save to local storage
-const checkedBox = (index) => {
-  finalTodo[index].completed = true;
-  storeItem();
-};
-
-// Set checkbox to false and save to local storage
-const notChecked = (index) => {
-  finalTodo[index].completed = false;
-  storeItem();
-};
 
 // Sort tasks in descending order based on their indexes
 const todoList = () => {
