+++ src/style.css
@@ -38,7 +38,7 @@
   justify-content: space-between;
 }
 
-.addtaks {
+.addtasks {
   font-style: italic;
   font-family: 'Roboto', sans-serif;
   font-size: 20px;
